Add a new user registration form with multiple steps and data validation
- Create a new user registration form with two steps:
  - Step 1: Enter basic account information (email and password).
  - Step 2: Enter personal user details (first name, last name, username, and country of residence).
- Add data validation using the `validate` function:
  - Check if the email is valid using a regular expression.
  - Ensure the password is strong, containing at least one uppercase letter, one number, and a minimum length of 6 characters.
  - Verify that the password matches the confirm password field.
  - Show error messages if the input is not valid.
- Add a loading state while sending data to the server.
- Send the data to an API to check if the user already exists and register a new account.
