+++ fakeDB/db.json
@@ -1,9 +1,20 @@
 {
   "users": [
     {
-      "username": "مرح",
-      "email": "mara@",
-      "password": "1233654789"
+      "email": "samasaadeh@gmail.com",
+      "password": "123456789",
+      "firstName": "Sma",
+      "lastName": "Saadeh",
+      "username": "Sma Saadeh",
+      "country": "فلسطين"
+    },
+    {
+      "email": "test@gmail.com",
+      "password": "$2b$10$J1a3ZdQ9/iuerOuWkc3upO.pt7Y76dpeS2Ym87KsME/kJo834XnIy",
+      "firstName": "test",
+      "lastName": "test",
+      "username": "test test",
+      "country": "تونس"
     }
   ],
   "courses": [
