+++ package.json
@@ -12,6 +12,7 @@
     "@chakra-ui/icons": "^2.2.4",
     "@chakra-ui/next-js": "^2.4.2",
     "@chakra-ui/utils": "^2.2.2",
+    "bcrypt": "^5.1.1",
     "json-server": "^1.0.0-beta.3",
     "next": "^15.0.2",
     "react": "^18",
